fix: Corregi los comentarios realizados en el Pr anterior
