+++ frontend/src/index.css
@@ -1,1 +1,9 @@
 @import 'tailwindcss';
+
+@theme {
+    --color-main: #fcf8fc;
+    --color-form: #f5eaf7;
+    --color-button-primary: #eabaff;
+}
+
+
