+++ frontend/src/pages/LoginPage.tsx
@@ -1,0 +1,5 @@
+import { Login } from '../components/Login/Login.tsx';
+
+export const LoginPage = () => (
+  <Login />
+);
